New RESTful endpoints added for Account and Bank

+++ backend-web/src/main/kotlin/com/kefir/web/DTOs/AccountResponse.kt
@@ -9,6 +9,6 @@ data class AccountResponse(val id: Long,
                            val bank: Long,
                            val CBU: String,
                            val balance: BigDecimal,
-                           val status: Int)
+                           val status: Long)
 
 fun Account.toResponse() = AccountResponse(id = this.id, customer = this.customer, currency = this.currency, bank = this.bank, CBU = this.CBU, balance = this.balance, status = this.status)
